agent_cloud_function/requirements.txt
google-cloud-secret-manager
google-cloud-aiplatform